# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restomate API",
        "description": "API de réservation de tables de restaurant. Gérez restaurants, tables et réservations avec authentification Sanctum.",
        "contact": {
            "name": "Alex AGO",
            "email": "alexyessougnonago@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://restomate-api.test/api/v1",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
        "/admin/users": {
            "get": {
                "tags": [
                    "Administration"
                ],
                "summary": "Liste tous les clients (Admin)",
                "operationId": "b55dcb1537fb85be8bae8aeee5b46fb1",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Alex"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateurs récupérés"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{id}": {
            "get": {
                "tags": [
                    "Administration"
                ],
                "summary": "Détail d'un client avec historique (Admin)",
                "operationId": "95f5fae091d1dece2b09f2afbac0cd74",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur récupéré"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{id}/toggle": {
            "patch": {
                "tags": [
                    "Administration"
                ],
                "summary": "Activer ou désactiver un compte client (Admin)",
                "operationId": "f837795bf71136c8c48362e01152272d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compte activé ou désactivé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/stats": {
            "get": {
                "tags": [
                    "Administration"
                ],
                "summary": "Statistiques globales de la plateforme (Admin)",
                "operationId": "5cbb48cca954b44e471bf5c809bf1ef4",
                "responses": {
                    "200": {
                        "description": "Statistiques récupérées"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Inscrire un nouveau client",
                "description": "Crée un compte client et retourne un token d'authentification",
                "operationId": "f8f20d8e5fa3ee0d9ab537e9ac6206b5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Alex AGO"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "alex@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+22997000000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inscription réussie"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connecter un utilisateur",
                "description": "Retourne un token d'authentification valide",
                "operationId": "3d1aae0ea009e03fadebec819afd774e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "alex@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie"
                    },
                    "401": {
                        "description": "Identifiants incorrects"
                    },
                    "403": {
                        "description": "Compte désactivé"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnecter l'utilisateur",
                "operationId": "0f604af450f66d99a17c511d6067a1c9",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Profil de l'utilisateur connecté",
                "operationId": "1482ae868da3ac9ed15d4972dc1724ec",
                "responses": {
                    "200": {
                        "description": "Profil récupéré"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entre ton token Sanctum obtenu après login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}